Troubleshooting mushroom tree widget
If Mushroom Tree teleportation widget is open, action to teleport to Verdant Valley goes off. Checking if the tree is nearby, accessing tree widget and teleporting is too many actions for 1 tick. Need to make task queue, or make states independent of onGameTick, so method call would not be recalled if the state is the same as previous state.